additional commit, add character counter
