+++ code/src/components/ThoughtForm.js
@@ -13,6 +13,7 @@ const ThoughtForm = ({ newThought, onNewThoughtChange, onFormSubmit, onThoughtLi
           onChange={onNewThoughtChange}
           className="thought"
           placeholder="write something nice here..." />
+        <p>{140 - newThought.length} / 140</p>
         <button
           type="submit"
           className="submit-button"
